# openapi-atspaces.yaml
openapi: 3.0.3
info:
  title: At Spaces API
  version: 1.0.0
  description: |
    API contract for At Spaces (NestJS + Prisma + PostgreSQL + Redis).
    Roles: customer, vendor, admin. Access token in JSON, refresh token in HttpOnly cookie.
servers:
  - url: http://localhost:3000/api
  - url: https://<domain>/api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          items: { type: object }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        hasNext: { type: boolean }
    UserPublic:
      type: object
      properties:
        id: { type: integer, example: 123 }
        role: { type: string, enum: [customer, vendor, admin] }
        fullName: { type: string, example: "John Doe" }
    AuthLoginResponse:
      type: object
      properties:
        accessToken: { type: string, example: "jwt" }
        user:
          $ref: '#/components/schemas/UserPublic'
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
...
    AdminResetPasswordRequest:
      type: object
      required: [resetToken, newPassword, totpCode]
      properties:
        resetToken: { type: string }
        newPassword: { type: string, format: password }
        totpCode: { type: string, example: "123456" }
    AdminLoginRequest:
      type: object
      required: [email, password, captchaToken]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        captchaToken: { type: string }
    AdminLoginResponse:
      type: object
      properties:
        preAuthToken: { type: string }
        mfaRequired: { type: boolean, example: true }
    AdminMfaVerifyRequest:
      type: object
      required: [preAuthToken, totpCode]
      properties:
        preAuthToken: { type: string }
        totpCode: { type: string, example: "123456" }
...
    BranchListItem:
      type: object
      properties:
        id: { type: integer, example: 10 }
        name: { type: string, example: "Branch A" }
        city: { type: string, example: "Amman" }
        address: { type: string, example: "..." }
...
    RejectReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 10 }
    UpdateStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [active, suspended] }
    ReportExportRequest:
      type: object
      required: [reportType, format, filters]
      properties:
        reportType: { type: string, example: revenue }
        format: { type: string, enum: [csv, pdf, xlsx] }
        filters:
          type: object
          additionalProperties: true
    ReportExportResponse:
      type: object
      properties:
        url: { type: string, example: "https://s3-presigned-url" }
        expiresIn: { type: integer, example: 300 }
...
paths:
...
  /admin/auth/login:
    post:
      tags: [Auth - Admin]
      summary: Admin login step 1 (password + captcha)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminLoginRequest' }
      responses:
        '200':
          description: Pre-auth token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminLoginResponse' }
  /admin/auth/mfa/verify:
    post:
      tags: [Auth - Admin]
      summary: Admin MFA verify step 2 (TOTP)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminMfaVerifyRequest' }
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              description: Refresh cookie (HttpOnly)
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthLoginResponse' }
...
  /admin/auth/forgot-password:
    post:
      tags: [Auth - Admin]
      summary: Admin forgot password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /admin/auth/reset-password:
    post:
      tags: [Auth - Admin]
      summary: Admin reset password (requires TOTP)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminResetPasswordRequest' }
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
...
  /branches:
    get:
      tags: [Branches]
      summary: List branches (filters + pagination)
      parameters:
        - name: city
          in: query
          required: false
          schema: { type: string }
        - name: serviceId
          in: query
          required: false
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged branches
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/BranchListItem' }
...
  /admin/approval-requests:
    get:
      tags: [Admin]
      summary: List approval requests
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [pending, approved, rejected] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged requests
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedResponse' }
  /admin/approval-requests/{id}:
    get:
      tags: [Admin]
      summary: Approval request details (includes integrity hash)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Details
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /admin/approval-requests/{id}/approve:
    post:
      tags: [Admin]
      summary: Approve request
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Approved
  /admin/approval-requests/{id}/reject:
    post:
      tags: [Admin]
      summary: Reject request (reason required)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectReasonRequest' }
      responses:
        '200':
          description: Rejected
  /admin/branches:
    get:
      tags: [Admin]
      summary: List branches
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged branches
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedResponse' }
  /admin/branches/{id}/status:
    patch:
      tags: [Admin]
      summary: Update branch status (audit logged)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateStatusRequest' }
      responses:
        '200':
          description: Updated
  /admin/vendors:
    get:
      tags: [Admin]
      summary: List vendors
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged vendors
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedResponse' }
  /admin/vendors/{id}/status:
    patch:
      tags: [Admin]
      summary: Update vendor status
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateStatusRequest' }
      responses:
        '200':
          description: Updated
  /admin/analytics:
    get:
      tags: [Admin]
      summary: Analytics overview (aggregated)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Analytics
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /admin/reports/export:
    post:
      tags: [Admin]
      summary: Export report (S3 presigned URL)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportExportRequest' }
      responses:
        '200':
          description: Presigned URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportExportResponse' }
  /admin/audit-log:
    get:
      tags: [Admin]
      summary: Audit log (paged)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: false
          schema: { type: string, format: date }
        - name: to
          in: query
          required: false
          schema: { type: string, format: date }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Paged audit log
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedResponse' }
...